requirements.txt
openai-whisper
ffmpeg-python
tk
